#12 Create DB tables according to class diagram
Added tables with sample data according to ER diagram. Few notes:

a) I could not go with table name "user" because this is a reserved word in postgresql. Now it's a little weird that all tables are in singular except one is in plural. Should we rename users or go with plural names for tables?
b) Few-word column naming is different. Sometimes it's snake_case (full_name, registration_no etc.) and sometimes it's camelCase (dateTime, imageUrl etc.). I believe we should unify the style.
c) From the diagram I did not understand the data type of eventType. Should it be an enum? For now I just made it a VARCHAR to be more flexible since we do not know all the possible event types and they might change in the future.
d) We might need a time-zone aware timestamp but I am not sure about it at the moment
e) I believe we should add a "modtime" column to users, animal, organization tables since they might be modified and modtime would be good to track and help debug what happened
f) We will ammend enum values later, now I do not know them. More about them:
https://www.postgresql.org/docs/9.1/functions-enum.html
https://dzone.com/articles/when-you-should-and-should-not