Addition of distinctive fields to animal
I believe these will be needed sooner or later. Data-types could be different.

I think it's hard to put all the possible breeds of all the species, may be not that good for an ENUM as well but this could be done.

I left allergy as a VARCHAR because this is also very undefined. Could be any allergy, could be a few.