#12 Create DB tables according to class diagram (#17)
* #12 Create DB tables according to class diagram

Added tables with sample data according to ER diagram. Few notes:

a) I could not go with table name "user" because this is a reserved word in postgresql. Now it's a little weird that all tables are in singular except one is in plural. Should we rename users or go with plural names for tables?
b) Few-word column naming is different. Sometimes it's snake_case (full_name, registration_no etc.) and sometimes it's camelCase (dateTime, imageUrl etc.). I believe we should unify the style.
c) From the diagram I did not understand the data type of eventType. Should it be an enum? For now I just made it a VARCHAR to be more flexible since we do not know all the possible event types and they might change in the future.
d) We might need a time-zone aware timestamp but I am not sure about it at the moment
e) I believe we should add a "modtime" column to users, animal, organization tables since they might be modified and modtime would be good to track and help debug what happened
f) We will ammend enum values later, now I do not know them. More about them:
https://www.postgresql.org/docs/9.1/functions-enum.html
https://dzone.com/articles/when-you-should-and-should-not

* Change users table to app_user

* Change event type data-type from VARCHAR to ENUM

Additionally separate event_general and event_medical ENUM since we also have different tables for them

* Unify the naming style of DB columns to snake_case

* Add modtime to organization, app_user, animal

Make it auto-update according to https://dba.stackexchange.com/a/157096

* Typo addoption -> adoption

* Change role_type from Stranger to Guest

* Change from Black Mamba to Green Mamba

* Move moddatetime extension and table triggers to schema.sql

Triggers automatically update modtime with any update of the table.

It's part of the definition of the tables and their behavior. Thus, it should stay in schema.sql.

* Remove redundant space

* Separate full_name to 2 separate fields "name" and "surname"

* Change organization id from SMALLSERIAL to SERIAL

Also change the foreign keys referencing it to INTEGER from SMALLINT

* Rename modtime to mod_time

* Decompose address field into country, city, street, house and flat fields

* Typo in SQL insert

* Change Postgresql date-type TIMESTAMP to upper-case

* Addition of distinctive fields to animal

I believe these will be needed sooner or later. Data-types could be different.

I think it's hard to put all the possible breeds of all the species, may be not that good for an ENUM as well but this could be done.

I left allergy as a VARCHAR because this is also very undefined. Could be any allergy, could be a few.

* Add missing "," in CREATE TABLE

* Add values for newly added fields to animal table